requirements.txt
Django==1.8.7
-e git+https://github.com/pinax/symposion.git#egg=symposion
django-appconf==1.0.1
django-model-utils==2.4.0
django-reversion==1.10.1
django-sitetree==1.5.1
django-taggit==0.18.0
django-timezone-field==1.3
django-user-accounts==1.3.1
easy-thumbnails==2.3
html5lib==0.9999999
markdown==2.6.5
pytz==2015.7
pinax-theme-bootstrap==8.0.1
django-jet==1.0.6
raven
razorpay
